requirements.txt
datasets 
sentence_transformers 
pinecone-client 
spacy 
nltk